fix(core): establish verified core baseline with 15 passing canonical tests
- Dynamic source compilation in build.py for complete src tree
- Corrected AcoClient/GlobalResourceRequest signature to long amounts
- Hardened GameplayAdapter with fail-fast null Item validation
- Implemented graceful shutdown (.stop()) on AcoServer
- Restored live TCP socket PING/PONG integration test (AcoIntegrationTest)
- Aligned GameplayBoundaryTest, MigrationProtocolTest, SimpleInitTest to production APIs
- Baseline build report updated: 15/15 canonical tests pass; desktop E2E marked environment-blocked

+++ src/sc/aco/AcoClient.java
@@ -21,7 +21,7 @@ public class AcoClient {
         // Implementation by integration wrapper
     }
 
-    public void sendGlobalResourceRequest(String item, int amount, String opId, String opType) {
+    public void sendGlobalResourceRequest(String item, long amount, String opId, String opType) {
         try {
             // protocol: item|amount|opId|opType
             String payload = item + "|" + amount + "|" + opId + "|" + opType;
